Make `TnOption.create()` invoke its callback
The response coming back from this endpoint has a different shape than
the `TnOption.get()` - it is wrapped in a `<TnOptionOrderResponse>`
element. That element doesn't have any direct `TnOptionGroups` children,
and so this method was throwing when trying to iterate over an undefined
attribute (in `item.tnOptionGroups.forEach(...)`).

Adjusting the tests to use an XML response that matches the
[API documentation](https://dev.bandwidth.com/numbers/apiReference.html)
demonstrates the failure well: the tests time out, because the thrown
error prevents the callback from firing.

The [package documentation](https://github.com/Bandwidth/node-numbers#add-a-portoutpasscode)
suggests that the shape of this response should match `TnOption.get()`,
which we can achieve by just unwrapping the contents from that outermost
`<TnOptionOrderResponse>` element.